+++ Changelog.md
@@ -25,6 +25,7 @@ OpenCore Changelog
 - Added workaround for read-only errors on some X299 boards
 - Added support for `x86legacy` Secure Boot model
 - Added missing Secure Boot NVRAM variables required by 11.0
+- Added setting of `system-id` NVRAM variable
 
 #### v0.6.2
 - Updated builtin firmware versions for SMBIOS and the rest